@@ -1,83 +1,73 @@
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Xml.Linq;
 using NUnit.Framework;
 using Toolbelt.Blazor.I18nText.Test.Internals;
-using static Toolbelt.Blazor.I18nText.Test.Internals.Shell;
 using static Toolbelt.Diagnostics.XProcess;
 
-namespace Toolbelt.Blazor.I18nText.Test
+namespace Toolbelt.Blazor.I18nText.Test;
+
+[Parallelizable(ParallelScope.All)]
+public class BuildTest
 {
-    [Parallelizable(ParallelScope.Children)]
-    public class BuildTest
+    private static readonly IEnumerable<string> HostingModels = new[] {
+        "Client", "Host", "Server" };
+
+    private static readonly IEnumerable<string> Frameworks = new[] {
+        "net6.0", "net7.0" };
+
+    public static readonly IEnumerable<object[]> Projects =
+        from startupProjName in HostingModels
+        from framework in Frameworks
+        select new object[] { startupProjName, framework };
+
+    [Test, TestCaseSource(nameof(Projects))]
+    public async Task BasicBuildTest(string startupProjName, string framework)
     {
-        public static IEnumerable<object[]> Projects = new[] {
-            // startupProjName
-            new object[]{ "Client" },
-            new object[]{ "Host" },
-            new object[]{ "Server" },
-        };
+        using var workSpace = new WorkSpace(startupProjName, framework, configuration: "Debug");
+        var distDir = Path.Combine(workSpace.OutputDir, Path.Combine("wwwroot/_content/i18ntext".Split('/')));
 
-        [Test]
-        [TestCaseSource(nameof(Projects))]
-        public async Task BasicBuildTest(string startupProjName)
-        {
-            using var workSpace = WorkSpace.Create(startupProjName);
-            var framework = workSpace.GetTargetFrameworkOfStartupProj();
-            var distDir = Path.Combine(workSpace.Bin, Path.Combine($"Debug/{framework}/wwwroot/_content/i18ntext".Split('/')));
+        await Start("dotnet", $"build -f:{framework}", workSpace.StartupProj).ExitCodeIs(0);
 
-            using var buildProcess = await Start("dotnet", "build", workSpace.StartupProj).WaitForExitAsync();
-            buildProcess.ExitCode.Is(0, message: buildProcess.Output);
+        var textResJsonFileNames = Directory.GetFiles(distDir, "*.*")
+            .Select(path => Path.GetFileName(path))
+            .OrderBy(name => name);
 
-            var textResJsonFileNames = Directory.GetFiles(distDir, "*.*")
-                .Select(path => Path.GetFileName(path))
-                .OrderBy(name => name);
+        textResJsonFileNames.Is(
+            "Lib4PackRef.I18nText.Text.en.json",
+            "Lib4PackRef.I18nText.Text.ja.json",
+            "Lib4PackRef6.I18nText.Text.en.json",
+            "Lib4PackRef6.I18nText.Text.ja.json",
+            "Lib4ProjRef.I18nText.Text.en.json",
+            "Lib4ProjRef.I18nText.Text.ja.json",
+            "SampleSite.Components.I18nText.Text.en.json",
+            "SampleSite.Components.I18nText.Text.ja.json"
+        );
+    }
 
-            textResJsonFileNames.Is(
-                "Lib4PackRef.I18nText.Text.en.json",
-                "Lib4PackRef.I18nText.Text.ja.json",
-                "Lib4PackRef6.I18nText.Text.en.json",
-                "Lib4PackRef6.I18nText.Text.ja.json",
-                "Lib4ProjRef.I18nText.Text.en.json",
-                "Lib4ProjRef.I18nText.Text.ja.json",
-                "SampleSite.Components.I18nText.Text.en.json",
-                "SampleSite.Components.I18nText.Text.ja.json"
-            );
-        }
+    [Test, TestCaseSource(nameof(Projects))]
+    public async Task PublishTest(string startupProjName, string framework)
+    {
+        using var workSpace = new WorkSpace(startupProjName, framework, configuration: "Release");
+        var wwwrootContentDir = Path.Combine(workSpace.PublishDir, "wwwroot", "_content");
+        var i18nDistDir = Path.Combine(wwwrootContentDir, "i18ntext");
+        var staticWebAssetDir = Path.Combine(wwwrootContentDir, "Toolbelt.Blazor.I18nText");
 
-        [Test]
-        [TestCaseSource(nameof(Projects))]
-        public async Task PublishTest(string startupProjName)
-        {
-            using var workSpace = WorkSpace.Create(startupProjName);
-            var framewrok = workSpace.GetTargetFrameworkOfStartupProj();
-            var publishDir = Path.Combine(workSpace.Bin, Path.Combine($"Release/{framewrok}/publish/".Split('/')));
-            var wwwrootContentDir = Path.Combine(publishDir, Path.Combine($"wwwroot/_content".Split('/')));
-            var i18nDistDir = Path.Combine(wwwrootContentDir, "i18ntext");
-            var staticWebAssetDir = Path.Combine(wwwrootContentDir, "Toolbelt.Blazor.I18nText");
+        await Start("dotnet", $"publish -c:Release -f:{framework} -p:UsingBrowserRuntimeWorkload=false --nologo", workSpace.StartupProj).ExitCodeIs(0);
 
-            using var publishProcess = await Start("dotnet", "publish -c:Release", workSpace.StartupProj).WaitForExitAsync();
-            publishProcess.ExitCode.Is(0, message: publishProcess.Output);
+        // Support client JavaScript file should be published into "_content/{PackageId}" folder.
+        FileIO.ExistsAnyFilesInDir(staticWebAssetDir, "script.min.js").IsTrue();
 
-            // Support client JavaScript file should be published into "_content/{PackageId}" folder.
-            Exists(staticWebAssetDir, "script.min.js").IsTrue();
+        var textResJsonFileNames = Directory.GetFiles(i18nDistDir, "*.json")
+            .Select(path => Path.GetFileName(path))
+            .OrderBy(name => name);
 
-            var textResJsonFileNames = Directory.GetFiles(i18nDistDir, "*.json")
-                .Select(path => Path.GetFileName(path))
-                .OrderBy(name => name);
-
-            textResJsonFileNames.Is(
-                "Lib4PackRef.I18nText.Text.en.json",
-                "Lib4PackRef.I18nText.Text.ja.json",
-                "Lib4PackRef6.I18nText.Text.en.json",
-                "Lib4PackRef6.I18nText.Text.ja.json",
-                "Lib4ProjRef.I18nText.Text.en.json",
-                "Lib4ProjRef.I18nText.Text.ja.json",
-                "SampleSite.Components.I18nText.Text.en.json",
-                "SampleSite.Components.I18nText.Text.ja.json"
-            );
-        }
+        textResJsonFileNames.Is(
+            "Lib4PackRef.I18nText.Text.en.json",
+            "Lib4PackRef.I18nText.Text.ja.json",
+            "Lib4PackRef6.I18nText.Text.en.json",
+            "Lib4PackRef6.I18nText.Text.ja.json",
+            "Lib4ProjRef.I18nText.Text.en.json",
+            "Lib4ProjRef.I18nText.Text.ja.json",
+            "SampleSite.Components.I18nText.Text.en.json",
+            "SampleSite.Components.I18nText.Text.ja.json"
+        );
     }
 }
